+++ src/main/java/com/frontApp/rateCalculator/CapitalizingRateCalculator.java
@@ -2,7 +2,7 @@ package com.frontApp.rateCalculator;
 
 public class CapitalizingRateCalculator implements InterestRateCalculator{
 	@Override
-	public double calculate(int years, double rate, int startSum) {
+	public double calculate(int years, double rate, double startSum) {
 		for(int i =0;i<years;i++){
 			startSum+=startSum*rate;
 		}
