+++ src/main/java/com/frontApp/rateCalculator/InterestRateCalculator.java
@@ -1,5 +1,5 @@
 package com.frontApp.rateCalculator;
 
 public interface InterestRateCalculator {
-	double calculate(int years,double rate,int startSum);
+	double calculate(int years,double rate,double startSum);
 }
