+++ src/main/java/com/frontApp/rateCalculator/NonCapitalizingRateCalculator.java
@@ -3,7 +3,7 @@ package com.frontApp.rateCalculator;
 
 public class NonCapitalizingRateCalculator implements InterestRateCalculator {
 	@Override
-	public double calculate(int years, double rate, int startSum) {
+	public double calculate(int years, double rate, double startSum) {
 		double resultSum=startSum;
 		for (int i = 0; i < years; i++) {
 			resultSum+=startSum*rate;
